src/requirements.txt
camel-tools
pyarabic
pandas
google-genai
google-generativeai
scikit-learn>=1.3.0
rank-bm25>=0.2.2
numpy>=1.24.0
